Reorg code, add tests to ensure all exposed things have types
