+++ homeassistant/components/google_assistant/smart_home.py
@@ -1,243 +1,21 @@
 """Support for Google Assistant Smart Home API."""
-from asyncio import gather
-from collections.abc import Mapping
 from itertools import product
 import logging
 
 from homeassistant.util.decorator import Registry
 
-from homeassistant.core import callback
 from homeassistant.const import (
-    CLOUD_NEVER_EXPOSED_ENTITIES, CONF_NAME, STATE_UNAVAILABLE,
-    ATTR_SUPPORTED_FEATURES, ATTR_ENTITY_ID, ATTR_DEVICE_CLASS,
-)
-from homeassistant.components import (
-    camera,
-    climate,
-    cover,
-    fan,
-    group,
-    input_boolean,
-    light,
-    lock,
-    media_player,
-    scene,
-    script,
-    switch,
-    vacuum,
-)
+    CLOUD_NEVER_EXPOSED_ENTITIES, ATTR_ENTITY_ID)
 
-
-from . import trait
 from .const import (
-    TYPE_LIGHT, TYPE_LOCK, TYPE_SCENE, TYPE_SWITCH, TYPE_VACUUM,
-    TYPE_THERMOSTAT, TYPE_FAN, TYPE_CAMERA, TYPE_BLINDS, TYPE_GARAGE,
-    TYPE_OUTLET,
-    CONF_ALIASES, CONF_ROOM_HINT,
-    ERR_FUNCTION_NOT_SUPPORTED, ERR_PROTOCOL_ERROR, ERR_DEVICE_OFFLINE,
-    ERR_UNKNOWN_ERROR,
+    ERR_PROTOCOL_ERROR, ERR_DEVICE_OFFLINE, ERR_UNKNOWN_ERROR,
     EVENT_COMMAND_RECEIVED, EVENT_SYNC_RECEIVED, EVENT_QUERY_RECEIVED
 )
-from .helpers import SmartHomeError, RequestData
+from .helpers import RequestData, GoogleEntity
+from .error import SmartHomeError
 
 HANDLERS = Registry()
 _LOGGER = logging.getLogger(__name__)
-
-DOMAIN_TO_GOOGLE_TYPES = {
-    camera.DOMAIN: TYPE_CAMERA,
-    climate.DOMAIN: TYPE_THERMOSTAT,
-    cover.DOMAIN: TYPE_BLINDS,
-    fan.DOMAIN: TYPE_FAN,
-    group.DOMAIN: TYPE_SWITCH,
-    input_boolean.DOMAIN: TYPE_SWITCH,
-    light.DOMAIN: TYPE_LIGHT,
-    lock.DOMAIN: TYPE_LOCK,
-    media_player.DOMAIN: TYPE_SWITCH,
-    scene.DOMAIN: TYPE_SCENE,
-    script.DOMAIN: TYPE_SCENE,
-    switch.DOMAIN: TYPE_SWITCH,
-    vacuum.DOMAIN: TYPE_VACUUM,
-}
-
-DEVICE_CLASS_TO_GOOGLE_TYPES = {
-    (cover.DOMAIN, cover.DEVICE_CLASS_GARAGE): TYPE_GARAGE,
-    (switch.DOMAIN, switch.DEVICE_CLASS_SWITCH): TYPE_SWITCH,
-    (switch.DOMAIN, switch.DEVICE_CLASS_OUTLET): TYPE_OUTLET,
-}
-
-
-def deep_update(target, source):
-    """Update a nested dictionary with another nested dictionary."""
-    for key, value in source.items():
-        if isinstance(value, Mapping):
-            target[key] = deep_update(target.get(key, {}), value)
-        else:
-            target[key] = value
-    return target
-
-
-def get_google_type(domain, device_class):
-    """Google type based on domain and device class."""
-    typ = DEVICE_CLASS_TO_GOOGLE_TYPES.get((domain, device_class))
-
-    return typ if typ is not None else DOMAIN_TO_GOOGLE_TYPES.get(domain)
-
-
-class _GoogleEntity:
-    """Adaptation of Entity expressed in Google's terms."""
-
-    def __init__(self, hass, config, state):
-        self.hass = hass
-        self.config = config
-        self.state = state
-        self._traits = None
-
-    @property
-    def entity_id(self):
-        """Return entity ID."""
-        return self.state.entity_id
-
-    @callback
-    def traits(self):
-        """Return traits for entity."""
-        if self._traits is not None:
-            return self._traits
-
-        state = self.state
-        domain = state.domain
-        features = state.attributes.get(ATTR_SUPPORTED_FEATURES, 0)
-        device_class = state.attributes.get(ATTR_DEVICE_CLASS)
-
-        self._traits = [Trait(self.hass, state, self.config)
-                        for Trait in trait.TRAITS
-                        if Trait.supported(domain, features, device_class)]
-        return self._traits
-
-    async def sync_serialize(self):
-        """Serialize entity for a SYNC response.
-
-        https://developers.google.com/actions/smarthome/create-app#actiondevicessync
-        """
-        state = self.state
-
-        # When a state is unavailable, the attributes that describe
-        # capabilities will be stripped. For example, a light entity will miss
-        # the min/max mireds. Therefore they will be excluded from a sync.
-        if state.state == STATE_UNAVAILABLE:
-            return None
-
-        entity_config = self.config.entity_config.get(state.entity_id, {})
-        name = (entity_config.get(CONF_NAME) or state.name).strip()
-        domain = state.domain
-        device_class = state.attributes.get(ATTR_DEVICE_CLASS)
-
-        # If an empty string
-        if not name:
-            return None
-
-        traits = self.traits()
-
-        # Found no supported traits for this entity
-        if not traits:
-            return None
-
-        device_type = get_google_type(domain,
-                                      device_class)
-
-        # If no device type found, skip
-        if not device_type:
-            return None
-
-        device = {
-            'id': state.entity_id,
-            'name': {
-                'name': name
-            },
-            'attributes': {},
-            'traits': [trait.name for trait in traits],
-            'willReportState': False,
-            'type': device_type,
-        }
-
-        # use aliases
-        aliases = entity_config.get(CONF_ALIASES)
-        if aliases:
-            device['name']['nicknames'] = aliases
-
-        for trt in traits:
-            device['attributes'].update(trt.sync_attributes())
-
-        room = entity_config.get(CONF_ROOM_HINT)
-        if room:
-            device['roomHint'] = room
-            return device
-
-        dev_reg, ent_reg, area_reg = await gather(
-            self.hass.helpers.device_registry.async_get_registry(),
-            self.hass.helpers.entity_registry.async_get_registry(),
-            self.hass.helpers.area_registry.async_get_registry(),
-        )
-
-        entity_entry = ent_reg.async_get(state.entity_id)
-        if not (entity_entry and entity_entry.device_id):
-            return device
-
-        device_entry = dev_reg.devices.get(entity_entry.device_id)
-        if not (device_entry and device_entry.area_id):
-            return device
-
-        area_entry = area_reg.areas.get(device_entry.area_id)
-        if area_entry and area_entry.name:
-            device['roomHint'] = area_entry.name
-
-        return device
-
-    @callback
-    def query_serialize(self):
-        """Serialize entity for a QUERY response.
-
-        https://developers.google.com/actions/smarthome/create-app#actiondevicesquery
-        """
-        state = self.state
-
-        if state.state == STATE_UNAVAILABLE:
-            return {'online': False}
-
-        attrs = {'online': True}
-
-        for trt in self.traits():
-            deep_update(attrs, trt.query_attributes())
-
-        return attrs
-
-    async def execute(self, command, data, params):
-        """Execute a command.
-
-        https://developers.google.com/actions/smarthome/create-app#actiondevicesexecute
-        """
-        executed = False
-        for trt in self.traits():
-            if trt.can_execute(command, params):
-                await trt.execute(command, data, params)
-                executed = True
-                break
-
-        if not executed:
-            raise SmartHomeError(
-                ERR_FUNCTION_NOT_SUPPORTED,
-                'Unable to execute {} for {}'.format(command,
-                                                     self.state.entity_id))
-
-    @callback
-    def async_update(self):
-        """Update the entity with latest info from Home Assistant."""
-        self.state = self.hass.states.get(self.entity_id)
-
-        if self._traits is None:
-            return
-
-        for trt in self._traits:
-            trt.state = self.state
 
 
 async def async_handle_message(hass, config, user_id, message):
@@ -311,7 +89,7 @@ async def async_devices_sync(hass, data, payload):
         if not data.config.should_expose(state):
             continue
 
-        entity = _GoogleEntity(hass, data.config, state)
+        entity = GoogleEntity(hass, data.config, state)
         serialized = await entity.sync_serialize()
 
         if serialized is None:
@@ -352,7 +130,7 @@ async def async_devices_query(hass, data, payload):
             devices[devid] = {'online': False}
             continue
 
-        entity = _GoogleEntity(hass, data.config, state)
+        entity = GoogleEntity(hass, data.config, state)
         devices[devid] = entity.query_serialize()
 
     return {'devices': devices}
@@ -396,7 +174,7 @@ async def handle_devices_execute(hass, data, payload):
                     }
                     continue
 
-                entities[entity_id] = _GoogleEntity(hass, data.config, state)
+                entities[entity_id] = GoogleEntity(hass, data.config, state)
 
             try:
                 await entities[entity_id].execute(execution['command'],
